requirements.txt
langchain
streamlit
langchain-groq
python-dotenv
fastapi
uvicorn